Add files via upload
These are scripts written by Junjian Wang, a co-mentor for this project. I have analyzed and annotated them to get the idea of how neural networks work. They use genotypes stored in the txt.raw file to predict phenotypes. This prediction is, then, compared to the true phenotypes in phen.csv. The loss, difference between the predicted and true outputs, is calculated and the model is adjusted to minimize this value. The txt.raw and phen.csv files could not be attached due to their large size.