Add files via upload
This script is a remodel of the previous models in sample_code using a transformer model. Similarly, it uses genotypes in txt.raw to predict phenotypes and comparing them to the true value found in phen.csv. The GPU memory is exceeded if the full dataset is put through, therefore, I adjusted the reading of the txt.raw file so that every 5 columns are used instead of the whole dataset. The txt.raw and phen.csv files aren't uploaded due to the txt.raw file being too large.