Add files via upload
These are the transformer models used to predict TF (transcription factors), NBA (number of piglets born alive), NBD (number of piglets born dead), and NW (number of piglets weaned. The model architecture and RandomizedSearchCV are the same for each model, however, the hyper-parameters differ based on the results of the RandomizedSearchCV. The files used for the genotypic and phenotypic data are not linked due to size constraints.